More cleanups (Q_NULLPTR and fixed a cast size warning)

+++ src/core/StelOpenGLArray.hpp
@@ -138,7 +138,7 @@ private:
 	int m_indexCount;
 	size_t m_memoryUsage;
 
-	qint32 m_offsets[ATTLOC_SIZE];
+	size_t m_offsets[ATTLOC_SIZE];
 	GLint m_sizes[ATTLOC_SIZE];
 	GLenum m_types[ATTLOC_SIZE];
 	GLsizei m_strides[ATTLOC_SIZE];
